move as_completed tests to new file
