+++ distributed/tests/test_client.py
@@ -372,44 +372,6 @@ def test_wait(c, s, a, b):
     assert x.status == y.status == 'finished'
 
 
-@gen_cluster(client=True)
-def test__as_completed(c, s, a, b):
-    x = c.submit(inc, 1)
-    y = c.submit(inc, 1)
-    z = c.submit(inc, 2)
-
-    from distributed.compatibility import Queue
-    queue = Queue()
-    yield _as_completed([x, y, z], queue)
-
-    assert queue.qsize() == 3
-    assert {queue.get(), queue.get(), queue.get()} == {x, y, z}
-
-    result = yield _first_completed([x, y, z])
-    assert result in [x, y, z]
-
-
-def test_as_completed(loop):
-    with cluster() as (s, [a, b]):
-        with Client(('127.0.0.1', s['port']), loop=loop) as c:
-            x = c.submit(inc, 1)
-            y = c.submit(inc, 2)
-            z = c.submit(inc, 1)
-
-            seq = as_completed([x, y, z])
-            assert isinstance(seq, Iterator)
-            assert set(seq) == {x, y, z}
-
-            assert list(as_completed([])) == []
-
-
-def test_as_completed_with_non_futures(loop):
-    with cluster() as (s, [a, b]):
-        with Client(('127.0.0.1', s['port']), loop=loop) as c:
-            with pytest.raises(TypeError):
-                list(as_completed([1, 2, 3]))
-
-
 def test_wait_sync(loop):
     with cluster() as (s, [a, b]):
         with Client(('127.0.0.1', s['port']), loop=loop) as c:
